requirements.txt

streamlit == 1.28.1

Pillow == 10.1.0
numpy == 1.26.1
